refactor(frontend): хуки, SCSS-модули, реорганизация страниц по папкам
- Выделены общие утилиты: компонент FlagIcon, formatBytes, buildChartOptions/buildChartData
- Логика страниц вынесена в кастомные хуки:
  useLogin, useNodes, useNodeDetail, useProxies, useProxyDetail
- Инлайн-стили и глобальные CSS-классы заменены SCSS-модулями
  для каждой страницы и компонента:
  Login, Nodes, NodeDetail, Proxies, ProxyDetail, Settings, ProxyCard
- App.css сокращён до базовых глобальных стилей
- Каждая страница перенесена в отдельную папку (Login/, Nodes/ и т.д.)
  с co-located .tsx и .module.scss файлами
- В каждой папке страницы добавлен barrel-файл index.ts для чистых импортов
