+++ frontend/package.json
@@ -26,6 +26,7 @@
     "@types/react": "^18.2.48",
     "@types/react-dom": "^18.2.18",
     "@vitejs/plugin-react": "^4.2.1",
+    "sass": "^1.99.0",
     "typescript": "^5.3.3",
     "vite": "^5.0.12"
   }
