+++ frontend/src/components/ProxyCard.tsx
@@ -2,14 +2,8 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Card, Button, Label, DropdownMenu } from '@gravity-ui/uikit';
 import { pauseProxy, unpauseProxy, ProxyData } from '../api';
-
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return '0 B';
-  const k = 1024;
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(1)) + ' ' + sizes[i];
-}
+import { formatBytes } from '../utils/format';
+import s from './ProxyCard.module.scss';
 
 interface Props {
   proxy: ProxyData;
@@ -66,47 +60,44 @@ export default function ProxyCard({ proxy, nodeId, nodeName, copied, onEdit, onD
   ];
 
   return (
-    <Card type="action" view="outlined" style={{ padding: 20, cursor: 'pointer' }} onClick={handleCardClick}>
-      <div className="proxy-card-header">
-        <span style={{ fontWeight: 600, fontSize: 15 }}>{proxy.name || `Proxy ${proxy.id}`}</span>
+    <Card type="action" view="outlined" className={s.card} onClick={handleCardClick}>
+      <div className={s.header}>
+        <span className={s.name}>{proxy.name || `Proxy ${proxy.id}`}</span>
         <Label theme={statusTheme} size="s">
           {statusLabel}
         </Label>
       </div>
 
       {proxy.note && (
-        <div style={{ fontSize: 13, color: 'var(--g-color-text-secondary)', marginBottom: 8 }}>
-          {proxy.note}
-        </div>
+        <div className={s.note}>{proxy.note}</div>
       )}
 
       {nodeName && (
-        <div className="proxy-card-field">
-          <span className="label">Нода</span>
+        <div className={s.field}>
+          <span className={s.label}>Нода</span>
           <span>{nodeName}</span>
         </div>
       )}
-      <div className="proxy-card-field">
-        <span className="label">Домен</span>
+      <div className={s.field}>
+        <span className={s.label}>Домен</span>
         <span>{proxy.domain}</span>
       </div>
-      <div className="proxy-card-field">
-        <span className="label">Трафик ↑</span>
+      <div className={s.field}>
+        <span className={s.label}>Трафик ↑</span>
         <span>{formatBytes(proxy.trafficUp || 0)}</span>
       </div>
-      <div className="proxy-card-field">
-        <span className="label">Трафик ↓</span>
+      <div className={s.field}>
+        <span className={s.label}>Трафик ↓</span>
         <span>{formatBytes(proxy.trafficDown || 0)}</span>
       </div>
       {proxy.connectedIps && proxy.connectedIps.length > 0 && (
-        <div className="proxy-card-field">
-          <span className="label">Подключения</span>
+        <div className={s.field}>
+          <span className={s.label}>Подключения</span>
           <span>{proxy.connectedIps.length}</span>
         </div>
       )}
 
-      {/* Футер: слева — DropdownMenu, справа — Ссылка */}
-      <div className="proxy-card-actions" style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+      <div className={s.actions}>
         <span onClick={(e) => e.stopPropagation()}>
           <DropdownMenu
             size="s"
