+++ frontend/src/pages/NodeDetail/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NodeDetail';
