+++ frontend/src/pages/Proxies/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Proxies';
