+++ frontend/src/pages/ProxyDetail/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProxyDetail';
